AddedTilemapEditorPackage
I added a tilemap editor package to the project. this tilemap editor can be used by the designers to difine where towers can or cannot be placed. the tower placer code can then test what tile is at the place where the player wants to place the tower and make a decision on whether the tower is placed or not.