Improved projectile<->enemy collision
I changed the onTriggerEnter function in the EnemyBase script, moving it to the projectile and activating a function in the enemy.
I also added pierce to the projectile, so the projectile only effects [pierce] enemy at a time, and the [pierce] value can be set for every projectile. I also added layers to the game and changed the prefabs, so they are set to the correct layer.